feat(styles): add theme and snippet style management functions
